requirements.txt
Flask==2.2.2
Flask_BasicAuth==0.2.0
icalendar==5.0.1
python_dateutil==2.8.2
pytz==2022.7.1
recurring_ical_events==2.0.0
requests==2.28.1
schedule==1.1.0
